% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_d2h.R
\name{check_model_data}
\alias{check_model_data}
\title{Check if model data exists locally}
\usage{
check_model_data(model_name, verbose = TRUE)
}
\arguments{
\item{model_name}{Name of the model}

\item{verbose}{Whether to print messages}
}
\value{
Logical indicating if data exists
}
\description{
Check if model data exists locally
}
\keyword{internal}
