% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_d2h.R
\name{detect_model_capabilities}
\alias{detect_model_capabilities}
\title{Detect model capabilities from static metadata}
\usage{
detect_model_capabilities(model_name)
}
\arguments{
\item{model_name}{Name of the model}
}
\value{
List of capability flags
}
\description{
Detect model capabilities from static metadata
}
