% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_with_progress}
\alias{download_with_progress}
\title{Download file with progress bar}
\usage{
download_with_progress(url, destfile, verbose = TRUE)
}
\arguments{
\item{url}{URL to download from}

\item{destfile}{Destination file path}

\item{verbose}{Whether to show progress bar}
}
\value{
Logical indicating success
}
\description{
Downloads a file with a text progress bar showing download progress.
}
\keyword{internal}
