% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_posteriors.R
\name{generate_fibonacci_sphere}
\alias{generate_fibonacci_sphere}
\title{Generate Fibonacci sphere points}
\usage{
generate_fibonacci_sphere(n_points = 125)
}
\arguments{
\item{n_points}{Number of points to generate}
}
\value{
Matrix with columns "lon" and "lat" in degrees
}
\description{
Generates evenly distributed points on a sphere using the Fibonacci spiral method.
}
\keyword{internal}
