% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_compatibility.R
\name{generate_model_description}
\alias{generate_model_description}
\title{Generate human-readable model description}
\usage{
generate_model_description(model_name, capabilities)
}
\arguments{
\item{model_name}{Name of the model}

\item{capabilities}{Model capabilities list}
}
\value{
Character description
}
\description{
Generate human-readable model description
}
\keyword{internal}
