% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{get_all_model_metadata}
\alias{get_all_model_metadata}
\title{Get all model metadata}
\usage{
get_all_model_metadata()
}
\value{
Named list of available models with metadata
}
\description{
Returns a list of all 14 models with their descriptions and properties.
The models are based on the spatial leafwax hierarchical Bayesian models.
}
