% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_data_url}
\alias{get_data_url}
\title{Get data download URLs}
\usage{
get_data_url(model_name, version = "latest", data_type = c("posteriors"))
}
\arguments{
\item{model_name}{Character string specifying the model name}

\item{version}{Version tag (e.g., "v1.0.0" or "latest")}

\item{data_type}{Type of data (only "posteriors" is currently supported)}
}
\value{
List of download URLs and filenames
}
\description{
Constructs download URLs for the configured model-data release.
}
\examples{
\dontrun{
urls <- get_data_url("baseline_sp", "latest")
}
}
