% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_download_files}
\alias{get_download_files}
\title{Get list of files to download for a model}
\usage{
get_download_files(model_name, data_type)
}
\arguments{
\item{model_name}{Character string specifying the model name}

\item{data_type}{Type of data: "minimal", "standard", or "full"}
}
\value{
List of file information
}
\description{
Internal function to determine which files need to be downloaded
based on model name and data type.
}
\keyword{internal}
