% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{get_model_info}
\alias{get_model_info}
\title{Get model info}
\usage{
get_model_info(model_name)
}
\arguments{
\item{model_name}{Name of the model}
}
\value{
List with model metadata
}
\description{
Get model info
}
