% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_compatibility.R
\name{get_model_recommendations}
\alias{get_model_recommendations}
\title{Get model recommendations based on available data}
\usage{
get_model_recommendations(
  has_elevation = FALSE,
  has_c4 = FALSE,
  has_pft = FALSE,
  prefer_spatial = TRUE,
  available_models = NULL
)
}
\arguments{
\item{has_elevation}{Whether elevation data is available}

\item{has_c4}{Whether C4 vegetation data is available}

\item{has_pft}{Whether PFT data is available}

\item{prefer_spatial}{Whether to prefer spatial models}

\item{available_models}{Vector of available model names (optional)}
}
\value{
Ranked list of recommended models
}
\description{
Suggests the best model(s) to use given the available predictors.
}
