% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_d2h.R
\name{get_model_size_estimate}
\alias{get_model_size_estimate}
\title{Get estimated download size for a model}
\usage{
get_model_size_estimate(model_name)
}
\arguments{
\item{model_name}{Name of the model}
}
\value{
Estimated size in MB
}
\description{
Get estimated download size for a model
}
\keyword{internal}
