% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_url_config}
\alias{get_url_config}
\title{Get URL configuration}
\usage{
get_url_config()
}
\value{
List with URL configuration
}
\description{
Loads the URL configuration from the package data.
}
\keyword{internal}
