% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_d2h.R
\name{load_model_posteriors}
\alias{load_model_posteriors}
\title{Load model posteriors from package data}
\usage{
load_model_posteriors(model_name, auto_download = NULL)
}
\arguments{
\item{model_name}{Name of the model}

\item{auto_download}{Logical whether to auto-download if missing}
}
\value{
Posterior draws as a data frame
}
\description{
Load model posteriors from package data
}
