% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation.R
\name{pair_distances}
\alias{pair_distances}
\title{Pairwise Euclidean distances between two coordinate matrices.}
\usage{
pair_distances(a, b)
}
\arguments{
\item{a}{matrix(n_a, 2)}

\item{b}{matrix(n_b, 2)}
}
\value{
matrix(n_a, n_b)
}
\description{
Pairwise Euclidean distances between two coordinate matrices.
}
