% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation.R
\name{predict_spatial_mpp}
\alias{predict_spatial_mpp}
\title{Predict spatial random effect at new location using mPP}
\usage{
predict_spatial_mpp(
  coords_std,
  knot_coords,
  z_spatial_draws,
  ls_gp_draws,
  sigma_gp_draws
)
}
\arguments{
\item{coords_std}{Standardized coordinates (length 2 vector)}

\item{knot_coords}{Matrix of knot coordinates (n_knots x 2)}

\item{z_spatial_draws}{Matrix of spatial random effects at knots (n_draws x n_knots)}

\item{ls_gp_draws}{Vector of GP length scale parameters (n_draws)}

\item{sigma_gp_draws}{Vector of GP variance parameters (n_draws)}
}
\value{
Vector of predicted spatial effects (n_draws)
}
\description{
Predict spatial random effect at new location using mPP
}
