% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\name{print.leafwax_lookup_table}
\alias{print.leafwax_lookup_table}
\title{Print method for lookup tables}
\usage{
\method{print}{leafwax_lookup_table}(x, ...)
}
\arguments{
\item{x}{A leafwax_lookup_table object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for lookup tables
}
