% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_posteriors.R
\name{print.leafwax_posterior}
\alias{print.leafwax_posterior}
\title{Print method for leafwax_posterior}
\usage{
\method{print}{leafwax_posterior}(x, ...)
}
\arguments{
\item{x}{A leafwax_posterior object}

\item{...}{Additional arguments}
}
\description{
Print method for leafwax_posterior
}
