% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{process_parallel}
\alias{process_parallel}
\title{Process chunks in parallel}
\usage{
process_parallel(data, chunks, model, n_cores, progress, ...)
}
\arguments{
\item{data}{Full dataset}

\item{chunks}{List of index vectors for chunks}

\item{model}{Model name}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}

\item{...}{Additional arguments}
}
\value{
List of results for each chunk
}
\description{
Process chunks in parallel
}
\keyword{internal}
