% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{process_sequential}
\alias{process_sequential}
\title{Process chunks sequentially with progress bar}
\usage{
process_sequential(data, chunks, model, progress, ...)
}
\arguments{
\item{data}{Full dataset}

\item{chunks}{List of index vectors for chunks}

\item{model}{Model name}

\item{progress}{Show progress bar}

\item{...}{Additional arguments}
}
\value{
List of results for each chunk
}
\description{
Process chunks sequentially with progress bar
}
\keyword{internal}
