% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R, R/model_utils.R
\name{select_best_model}
\alias{select_best_model}
\title{Select best model based on available data}
\usage{
select_best_model(data)

select_best_model(data)
}
\arguments{
\item{data}{Data frame with predictor variables}

\item{has_elevation}{Logical, whether elevation data is available}

\item{has_c4}{Logical, whether C4 vegetation data is available}

\item{has_pft}{Logical, whether PFT data is available}

\item{prefer_spatial}{Logical, whether to prefer spatial models}

\item{verbose}{Logical, whether to print selection reasoning}
}
\value{
Character string with selected model name

Recommended model name
}
\description{
Automatically selects the most appropriate model based on which
covariates are available in the data.
}
