% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{setup_leafwax_data}
\alias{setup_leafwax_data}
\title{Setup leafwax data management}
\usage{
setup_leafwax_data(reset = FALSE)
}
\arguments{
\item{reset}{Logical, whether to reset to defaults}
}
\description{
Interactive setup wizard for configuring leafwax data management options.
}
\examples{
\dontrun{
# Run interactive setup
setup_leafwax_data()
}
}
