% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation.R
\name{standardize_coords}
\alias{standardize_coords}
\title{Standardize a coord matrix using the scaling parameters.}
\usage{
standardize_coords(coords, scaling)
}
\arguments{
\item{coords}{matrix(n, 2) of (lon, lat) in degrees}

\item{scaling}{list with $lon_mean, $lon_sd, $lat_mean, $lat_sd}
}
\description{
Standardize a coord matrix using the scaling parameters.
}
