% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_d2h.R
\name{use_example_data}
\alias{use_example_data}
\title{Use lightweight example data}
\usage{
use_example_data(model_name)
}
\arguments{
\item{model_name}{Name of the model}
}
\value{
Minimal posterior draws
}
\description{
Falls back to minimal example data when full data is not available
}
\keyword{internal}
