% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data for inversion}
\usage{
validate_inputs(
  d2h_wax,
  longitude,
  latitude,
  d2h_wax_err = NULL,
  elevation = NULL,
  c4_fraction = NULL,
  pft_tree = NULL,
  pft_shrub = NULL,
  pft_grass = NULL,
  model_name = "b0b1"
)
}
\arguments{
\item{d2h_wax}{Leaf wax d2H values}

\item{longitude}{Longitude values}

\item{latitude}{Latitude values}

\item{d2h_wax_err}{Measurement uncertainties}

\item{elevation}{Elevation values}

\item{c4_fraction}{C4 vegetation fraction}

\item{pft_tree}{Tree PFT fraction}

\item{pft_shrub}{Shrub PFT fraction}

\item{pft_grass}{Grass PFT fraction}

\item{model_name}{Name of model to use}
}
\value{
List of validated inputs
}
\description{
Checks that input data meets requirements for the specified model
and returns cleaned, validated data.
}
