% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\name{validate_lookup_table}
\alias{validate_lookup_table}
\title{Validate lookup table}
\usage{
validate_lookup_table(lookup_table)
}
\arguments{
\item{lookup_table}{Object to validate}
}
\value{
Logical TRUE if valid, otherwise throws an error
}
\description{
Checks that a lookup table is valid and contains all required components.
}
