% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_cache_files}
\alias{get_cache_files}
\title{Get cache files for a model}
\usage{
get_cache_files(model_name, data_type, cache_dir)
}
\arguments{
\item{model_name}{Model name}

\item{data_type}{Data type (NULL for all)}

\item{cache_dir}{Cache directory path}
}
\value{
Character vector of file paths
}
\description{
Internal function to get list of cache files for a model.
}
\keyword{internal}
