% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_data_manifest}
\alias{get_data_manifest}
\title{Get data manifest}
\usage{
get_data_manifest()
}
\value{
List with manifest data
}
\description{
Loads or downloads the data manifest with file checksums.
}
\keyword{internal}
