% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{select_best_model}
\alias{select_best_model}
\title{Select best model based on available data}
\usage{
select_best_model(data)
}
\arguments{
\item{data}{Data frame with predictor variables}
}
\value{
Recommended model name
}
\description{
Select best model based on available data
}
